Separate Flower client and server start utilities
