+++ src/flower/server/__init__.py
@@ -12,20 +12,4 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 # ==============================================================================
-"""Flower main package."""
-
-
-from .client import Client
-from .client_manager import SimpleClientManager
-from .history import History
-from .server import Server
-from .strategy.parameter import parameters_to_weights, weights_to_parameters
-from .typing import (
-    EvaluateIns,
-    EvaluateRes,
-    FitIns,
-    FitRes,
-    Parameters,
-    ParametersRes,
-    Weights,
-)
+"""Flower server app."""
