+++ src/flower_benchmark/tf_hotkey/server.py
@@ -138,7 +138,7 @@ def main() -> None:
 
     # Run server
     server = flwr.Server(client_manager=client_manager, strategy=strategy)
-    flwr.app.start_server(
+    flwr.server.start_server(
         DEFAULT_SERVER_ADDRESS, server, config={"num_rounds": server_setting.rounds},
     )
