+++ src/flower_example/pytorch/client.py
@@ -124,7 +124,7 @@ def main() -> None:
 
     # Start client
     client = CifarClient(args.cid, model, trainset, testset)
-    fl.app.start_client(args.server_address, client)
+    fl.client.start_client(args.server_address, client)
 
 
 if __name__ == "__main__":
