+++ src/flower_example/pytorch/server.py
@@ -86,7 +86,7 @@ def main() -> None:
     server = fl.Server(client_manager=client_manager, strategy=strategy)
 
     # Run server
-    fl.app.start_server(
+    fl.server.start_server(
         args.server_address, server, config={"num_rounds": args.rounds},
     )
